+++ Engine/Headers/engine/modeling/mesh.h
@@ -12,13 +12,13 @@ struct _mesh {
 	char* Name;
 	bool SmoothingEnabled;
 	// The number of vector3s in the VertexData array
-	size_t VertexCount;
+	ulong VertexCount;
 	vector3* Vertices;
 	// The number of vector2 in the TextureVertexData array
-	size_t TextureCount;
+	ulong TextureCount;
 	vector2* TextureVertices;
 	// The number of vector3 in the NormalVertexData array
-	size_t NormalCount;
+	ulong NormalCount;
 	vector3* NormalVertices;
 	char* MaterialName;
 };
